use new api method in azure monitor
